[IMP] Moved code out of util.py into blueprints namespace
