+++ source/app/blueprints/access_controls.py
@@ -21,26 +21,33 @@ import logging as log
 import uuid
 from functools import wraps
 
+import jwt
+import requests
+
 from flask import Request
 from flask import request
-from flask import session
 from flask import render_template
+from flask import session
 from flask_login import current_user
+from flask_login import login_user
 from flask_wtf import FlaskForm
+from jwt import PyJWKClient
+from requests.auth import HTTPBasicAuth
 from werkzeug.utils import redirect
 
 from app import TEMPLATE_PATH
+
 from app import app
+from app import db
 from app.datamgmt.case.case_db import get_case
 from app.datamgmt.manage.manage_access_control_db import user_has_client_access
+from app.datamgmt.manage.manage_users_db import get_user
 from app.iris_engine.access_control.utils import ac_fast_check_user_has_case_access
 from app.iris_engine.access_control.utils import ac_get_effective_permissions_of_user
+from app.iris_engine.utils.tracker import track_activity
+from app.models import Cases
 from app.models.authorization import Permissions
 from app.models.authorization import CaseAccessLevel
-from app.util import _local_authentication_process
-from app.util import _local_authentication_process
-from app.util import _local_authentication_process
-from app.util import _oidc_proxy_authentication_process
 
 from app.util import update_current_case
 from app.util import log_exception_and_error
@@ -360,6 +367,106 @@ def ac_api_requires_client_access():
             return f(*args, **kwargs)
         return wrap
     return inner_wrap
+
+
+def _authenticate_with_email(user_email):
+    user = get_user(user_email, id_key="email")
+    if not user:
+        log.error(f'User with email {user_email} is not registered in the IRIS')
+        return False
+
+    login_user(user)
+    track_activity(f"User '{user.id}' successfully logged-in", ctx_less=True)
+
+    caseid = user.ctx_case
+    session['permissions'] = ac_get_effective_permissions_of_user(user)
+
+    if caseid is None:
+        case = Cases.query.order_by(Cases.case_id).first()
+        user.ctx_case = case.case_id
+        user.ctx_human_case = case.name
+        db.session.commit()
+
+    session['current_case'] = {
+        'case_name': user.ctx_human_case,
+        'case_info': "",
+        'case_id': user.ctx_case
+    }
+
+    return True
+
+
+def _oidc_proxy_authentication_process(incoming_request: Request):
+    # Get the OIDC JWT authentication token from the request header
+    authentication_token = incoming_request.headers.get('X-Forwarded-Access-Token', '')
+
+    if app.config.get("AUTHENTICATION_TOKEN_VERIFY_MODE") == 'lazy':
+        user_email = incoming_request.headers.get('X-Email')
+
+        if user_email:
+            return _authenticate_with_email(user_email.split(',')[0])
+
+    elif app.config.get("AUTHENTICATION_TOKEN_VERIFY_MODE") == 'introspection':
+        # Use the authentication server's token introspection endpoint in order to determine if the request is valid /
+        # authenticated. The TLS_ROOT_CA is used to validate the authentication server's certificate.
+        # The other solution was to skip the certificate verification, BUT as the authentication server might be
+        # located on another server, this check is necessary.
+
+        introspection_body = {"token": authentication_token}
+        introspection = requests.post(
+            app.config.get("AUTHENTICATION_TOKEN_INTROSPECTION_URL"),
+            auth=HTTPBasicAuth(app.config.get('AUTHENTICATION_CLIENT_ID'), app.config.get('AUTHENTICATION_CLIENT_SECRET')),
+            data=introspection_body,
+            verify=app.config.get("TLS_ROOT_CA")
+        )
+        if introspection.status_code == 200:
+            response_json = introspection.json()
+
+            if response_json.get("active", False) is True:
+                user_email = response_json.get("sub")
+                return _authenticate_with_email(user_email=user_email)
+
+            else:
+                log.info("USER IS NOT AUTHENTICATED")
+                return False
+
+    elif app.config.get("AUTHENTICATION_TOKEN_VERIFY_MODE") == 'signature':
+        # Use the JWKS urls provided by the OIDC discovery to fetch the signing keys
+        # and check the signature of the token
+        try:
+            jwks_client = PyJWKClient(app.config.get("AUTHENTICATION_JWKS_URL"))
+            signing_key = jwks_client.get_signing_key_from_jwt(authentication_token)
+
+            try:
+
+                data = jwt.decode(
+                    authentication_token,
+                    signing_key.key,
+                    algorithms=["RS256"],
+                    audience=app.config.get("AUTHENTICATION_AUDIENCE"),
+                    options={"verify_exp": app.config.get("AUTHENTICATION_VERIFY_TOKEN_EXP")},
+                )
+
+            except jwt.ExpiredSignatureError:
+                log.error("Provided token has expired")
+                return False
+
+        except Exception as e:
+            log.error(f"Error decoding JWT. {e.__str__()}")
+            return False
+
+        # Extract the user email
+        user_email = data.get("sub")
+
+        return _authenticate_with_email(user_email)
+
+    else:
+        log.error("ERROR DURING TOKEN INTROSPECTION PROCESS")
+        return False
+
+
+def _local_authentication_process(incoming_request: Request):
+    return current_user.is_authenticated
 
 
 def is_user_authenticated(incoming_request: Request):
